requirements.txt
fastapi
uvicorn
cohere
qdrant-client
google-genai
python-dotenv
